homogeneise rule numeric and tranfer time related fct to datetime
