+++ cuallee/family/datetime/__init__.py
@@ -368,3 +368,118 @@ class DateTimeCheck(ABC):
             >> self._rule
         )
         return self
+
+    def is_after_than(
+        self,
+        column: str,
+        value: datetime,
+        pct: float = 1.0,
+    ):
+        """
+        Validation for date/timestamp greater than value
+
+        Args:
+            column (str): Column name in dataframe
+            value (datetime): The condition for the column to match
+            pct (float): The threshold percentage required to pass
+        """
+        (
+            Rule(
+                "is_greater_than",
+                column,
+                value,
+                [RuleDataType.DATE, RuleDataType.TIMESTAMP],
+                pct,
+                options={"name": "is_after_than"},
+            )
+            >> self._rule
+        )
+        return self
+
+    def is_after_or_equal_than(self, column: str, value: datetime, pct: float = 1.0):
+        """
+        Validation for datetime greater or equal than value
+
+        Args:
+            column (str): Column name in dataframe
+            value (datetime): The condition for the column to match
+            pct (float): The threshold percentage required to pass
+        """
+        (
+            Rule(
+                "is_greater_or_equal_than",
+                column,
+                value,
+                [RuleDataType.DATE, RuleDataType.TIMESTAMP],
+                pct,
+                options={"name": "is_after_or_equal_than"},
+            )
+            >> self._rule
+        )
+        return self
+
+    def is_before_than(self, column: str, value: datetime, pct: float = 1.0):
+        """
+        Validation for datetime less than value
+
+        Args:
+            column (str): Column name in dataframe
+            value (datetime): The condition for the column to match
+            pct (float): The threshold percentage required to pass
+        """
+        (
+            Rule(
+                "is_less_than",
+                column,
+                value,
+                [RuleDataType.DATE, RuleDataType.TIMESTAMP],
+                pct,
+                options={"name": "is_before_than"},
+            )
+            >> self._rule
+        )
+        return self
+
+    def is_before_or_equal_than(self, column: str, value: datetime, pct: float = 1.0):
+        """
+        Validation for datetime less or equal than value
+
+        Args:
+            column (str): Column name in dataframe
+            value (datetime): The condition for the column to match
+            pct (float): The threshold percentage required to pass
+        """
+        (
+            Rule(
+                "is_less_or_equal_than",
+                column,
+                value,
+                [RuleDataType.DATE, RuleDataType.TIMESTAMP],
+                pct,
+                options={"name": "is_before_or_equal_than"},
+            )
+            >> self._rule
+        )
+        return self
+
+    def is_on_date(self, column: str, value: datetime, pct: float = 1.0):
+        """
+        Validation for datetime column equal than value
+
+        Args:
+            column (str): Column name in dataframe
+            value (datetime): The condition for the column to match
+            pct (float): The threshold percentage required to pass
+        """
+        (
+            Rule(
+                "is_equal_than",
+                column,
+                value,
+                [RuleDataType.DATE, RuleDataType.TIMESTAMP],
+                pct,
+                options={"name": "is_on_date"},
+            )
+            >> self._rule
+        )
+        return self
