+++ cuallee/family/numeric/__init__.py
@@ -1,5 +1,4 @@
 from abc import ABC, abstractmethod
-from datetime import datetime
 
 from cuallee.core.rule import Rule, RuleDataType
 
@@ -92,32 +91,14 @@ class NumericCheck(ABC):
             column (str): Column name in dataframe
             pct (float): The threshold percentage required to pass
         """
-        return self.is_greater_than(
-            column, 0, RuleDataType.NUMERIC, pct, options={"name": "is_positive"}
-        )
-
-    def is_after_than(
-        self,
-        column: str,
-        value: datetime,
-        pct: float = 1.0,
-    ):
-        """
-        Validation for date/timestamp greater than value
-
-        Args:
-            column (str): Column name in dataframe
-            value (datetime): The condition for the column to match
-            pct (float): The threshold percentage required to pass
-        """
         (
             Rule(
                 "is_greater_than",
                 column,
-                value,
-                [RuleDataType.DATE, RuleDataType.TIMESTAMP],
+                0,
+                RuleDataType.NUMERIC,
                 pct,
-                options={"name": "is_after_than"},
+                options={"name": "is_positive"},
             )
             >> self._rule
         )
@@ -167,28 +148,6 @@ class NumericCheck(ABC):
         )
         return self
 
-    def is_after_or_equal_than(self, column: str, value: datetime, pct: float = 1.0):
-        """
-        Validation for datetime greater or equal than value
-
-        Args:
-            column (str): Column name in dataframe
-            value (datetime): The condition for the column to match
-            pct (float): The threshold percentage required to pass
-        """
-        (
-            Rule(
-                "is_greater_or_equal_than",
-                column,
-                value,
-                [RuleDataType.DATE, RuleDataType.TIMESTAMP],
-                pct,
-                options={"name": "is_after_or_equal_than"},
-            )
-            >> self._rule
-        )
-        return self
-
     def is_in_millions(self, column: str, pct: float = 1.0):
         """
         Validates that a column has values greater than 1M (1e6)
@@ -197,9 +156,18 @@ class NumericCheck(ABC):
             column (str): Column name in dataframe
             pct (float): The threshold percentage required to pass
         """
-        return self.is_greater_or_equal_than(
-            column, 1e6, RuleDataType.NUMERIC, pct, options={"name": "is_in_millions"}
+        (
+            Rule(
+                "is_greater_or_equal_than",
+                column,
+                1e6,
+                RuleDataType.NUMERIC,
+                pct,
+                options={"name": "is_in_millions"},
+            )
+            >> self._rule
         )
+        return self
 
     def is_in_billions(self, column: str, pct: float = 1.0):
         """
@@ -209,9 +177,18 @@ class NumericCheck(ABC):
             column (str): Column name in dataframe
             pct (float): The threshold percentage required to pass
         """
-        return self.is_greater_or_equal_than(
-            column, 1e9, RuleDataType.NUMERIC, pct, options={"name": "is_in_billions"}
+        (
+            Rule(
+                "is_greater_or_equal_than",
+                column,
+                1e9,
+                RuleDataType.NUMERIC,
+                pct,
+                options={"name": "is_in_billions"},
+            )
+            >> self._rule
         )
+        return self
 
     def is_less_than(self, column: str, value: float, pct: float = 1.0):
         """
@@ -288,27 +265,14 @@ class NumericCheck(ABC):
             column (str): Column name in dataframe
             pct (float): The threshold percentage required to pass
         """
-        return self.is_less_than(
-            column, 0, RuleDataType.NUMERIC, pct, options={"name": "is_negative"}
-        )
-
-    def is_before_than(self, column: str, value: datetime, pct: float = 1.0):
-        """
-        Validation for datetime less than value
-
-        Args:
-            column (str): Column name in dataframe
-            value (datetime): The condition for the column to match
-            pct (float): The threshold percentage required to pass
-        """
         (
             Rule(
                 "is_less_than",
                 column,
-                value,
-                [RuleDataType.DATE, RuleDataType.TIMESTAMP],
+                0,
+                RuleDataType.NUMERIC,
                 pct,
-                options={"name": "is_before_than"},
+                options={"name": "is_negative"},
             )
             >> self._rule
         )
@@ -358,28 +322,6 @@ class NumericCheck(ABC):
         )
         return self
 
-    def is_before_or_equal_than(self, column: str, value: datetime, pct: float = 1.0):
-        """
-        Validation for datetime less or equal than value
-
-        Args:
-            column (str): Column name in dataframe
-            value (datetime): The condition for the column to match
-            pct (float): The threshold percentage required to pass
-        """
-        (
-            Rule(
-                "is_less_or_equal_than",
-                column,
-                value,
-                [RuleDataType.DATE, RuleDataType.TIMESTAMP],
-                pct,
-                options={"name": "is_before_or_equal_than"},
-            )
-            >> self._rule
-        )
-        return self
-
     def is_equal_than(self, column: str, value: float, pct: float = 1.0):
         """
         Validation for numeric column equal than value
@@ -419,28 +361,6 @@ class NumericCheck(ABC):
                 [RuleDataType.NUMERIC, RuleDataType.DATE, RuleDataType.TIMESTAMP],
                 pct,
                 options={"name": "eq"},
-            )
-            >> self._rule
-        )
-        return self
-
-    def is_on_date(self, column: str, value: datetime, pct: float = 1.0):
-        """
-        Validation for datetime column equal than value
-
-        Args:
-            column (str): Column name in dataframe
-            value (datetime): The condition for the column to match
-            pct (float): The threshold percentage required to pass
-        """
-        (
-            Rule(
-                "is_equal_than",
-                column,
-                value,
-                [RuleDataType.DATE, RuleDataType.TIMESTAMP],
-                pct,
-                options={"name": "is_on_date"},
             )
             >> self._rule
         )
